+++ src/main.jsx
@@ -5,7 +5,7 @@ import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 
 import LandingPage from "./components/LandingPage/LandingPage.jsx";
 import AnalysisPage from "./components/AnalysisPage/AnalysisPage.jsx";
-import Login from "./components/loginRegister/login/Login.jsx";
+import LoginPage from "./components/loginRegister/LoginPage/LoginPage.jsx";
 import Register from "./components/loginRegister/register/Register.jsx";
 
 export const RootComponent = () => {
@@ -13,7 +13,7 @@ export const RootComponent = () => {
     <Routes>
       <Route path="/" element={<App />}>
         <Route index element={<LandingPage />} />
-        <Route path="login" element={<Login />} />
+        <Route path="login" element={<LoginPage />} />
         <Route path="register" element={<Register />} />
         <Route path="analysis" element={<AnalysisPage />} />
       </Route>
